pyproject.toml
[project]
name = "sea-battle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "advanced-alchemy>=1.8.0",
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "dotenv>=0.9.9",
    "fastapi>=0.121.0",
    "passlib>=1.7.4",
    "pillow>=12.0.0",
    "pyjwt>=2.10.1",
    "pydantic-settings>=2.11.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
]
